requirements.txt
pymongo
python-dotenv
requests
pandas
fastapi
uvicorn[standard]
